+++ secure-gate-core/src/macros/newtype_common.rs
@@ -287,3 +287,25 @@ macro_rules! __sg_newtype_len {
         }
     };
 }
+
+#[doc(hidden)]
+#[macro_export]
+#[cfg(feature = "encoding-base64")]
+macro_rules! __sg_if_base64 { ($($t:tt)*) => { $($t)* }; }
+#[doc(hidden)]
+#[macro_export]
+#[cfg(not(feature = "encoding-base64"))]
+macro_rules! __sg_if_base64 {
+    ($($t:tt)*) => {};
+}
+
+#[doc(hidden)]
+#[macro_export]
+#[cfg(feature = "encoding-bech32m")]
+macro_rules! __sg_if_bech32m { ($($t:tt)*) => { $($t)* }; }
+#[doc(hidden)]
+#[macro_export]
+#[cfg(not(feature = "encoding-bech32m"))]
+macro_rules! __sg_if_bech32m {
+    ($($t:tt)*) => {};
+}